+++ src/TestingApp.Domain/Entities/Quizes/QuizResult.cs
@@ -5,10 +5,10 @@ namespace TestingApp.Domain.Entities.Quizes
 {
     public class QuizResult : Auditable
     {
-        public int CorrectAnswers { get; set; }
-        public int QuizId { get; set; }
+        public long CorrectAnswers { get; set; }
+        public long QuizId { get; set; }
         public Quiz Quiz { get; set; }
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public User User { get; set; }
         public ICollection<SolvedQuestion> SolvedQuestions { get; set; }
     }
